+++ src/main/java/be/technifutur/gestionchampionnatfootrest/data/repo/JourneeRepository.java
@@ -10,5 +10,5 @@ import be.technifutur.gestionchampionnatfootrest.models.entities.Journee;
 public interface JourneeRepository extends JpaRepository<Journee,Long> {
     Optional<Journee> findByNumero(int numero);
     @Query(value = "select * from journee j where j.numero = ?1 and j.championnat_id = ?2", nativeQuery = true)
-    Optional<Journee> findByNumeroAndChampionnat(int numero, int championnat);
+    Optional<Journee> findByNumeroAndChampionnat(int numero, Long championnat);
 }
