+++ src/main/java/be/technifutur/gestionchampionnatfootrest/metiers/services/JourneeService.java
@@ -3,6 +3,6 @@ package be.technifutur.gestionchampionnatfootrest.metiers.services;
 import be.technifutur.gestionchampionnatfootrest.models.dtos.JourneeDTO;
 
 public interface JourneeService {
-    JourneeDTO getOneByNumeroAndChampionnat(int numero, int championnatId);
+    JourneeDTO getOneByNumeroAndChampionnat(int numero, Long championnatId);
     JourneeDTO getOneByNumero(int numero);
 }
